refactor round-2 better abstraction and eliminate lazy import
